* commit '719e2c478ed2a20d74890f25b9f503c355d69b03':
  Added conversion back to RobotSide.
  Fixed a test.
  Fixed up the ROS API.
  Updated scripts.
  Trying to trick bamboo to not fail because of weird timeout declaration.
  Added statics.
  Added test that enforce the declaration of static values for enums.
  Renamed the byte fields of all packets to follow the new convention.
  Added test to assert byte field naming convention.
  Replaced enum fields with byte fields in all packets.
  Changed QueueableMessage to have 1 setter/getter per field.
  Nuked ValkyrieLowLevelControlModeMessage.
  Added test for enum fields in Packets.

+++ ihmc-avatar-interfaces/src/main/java/us/ihmc/avatar/networkProcessor/footstepPlanningToolboxModule/FootstepDataListWithSwingOverTrajectoriesAssembler.java
@@ -80,7 +80,7 @@ public class FootstepDataListWithSwingOverTrajectoriesAssembler
          FootstepDataMessage footstepDataMessage = HumanoidMessageTools.createFootstepDataMessage(simpleFootstep.getRobotSide(), new Point3D(swingEndPose.getPosition()), new Quaternion(swingEndPose.getOrientation()));
 
          double maxSpeedDimensionless = swingOverPlanarRegionsTrajectoryExpander.expandTrajectoryOverPlanarRegions(stanceFootPose, swingStartPose, swingEndPose, planarRegionsList);
-         footstepDataMessage.setTrajectoryType(TrajectoryType.CUSTOM);
+         footstepDataMessage.setTrajectoryType(TrajectoryType.CUSTOM.toByte());
          Point3D[] waypoints = new Point3D[] {new Point3D(), new Point3D()};
          waypoints[0].set(swingOverPlanarRegionsTrajectoryExpander.getExpandedWaypoints().get(0));
          waypoints[1].set(swingOverPlanarRegionsTrajectoryExpander.getExpandedWaypoints().get(1));
@@ -142,7 +142,7 @@ public class FootstepDataListWithSwingOverTrajectoriesAssembler
          FootstepDataMessage footstepDataMessage = HumanoidMessageTools.createFootstepDataMessage(footstep.getRobotSide(), new Point3D(swingEndPose.getPosition()), new Quaternion(swingEndPose.getOrientation()));
 
          swingOverPlanarRegionsTrajectoryExpander.expandTrajectoryOverPlanarRegions(stanceFootPose, swingStartPose, swingEndPose, planarRegionsList);
-         footstepDataMessage.setTrajectoryType(TrajectoryType.CUSTOM);
+         footstepDataMessage.setTrajectoryType(TrajectoryType.CUSTOM.toByte());
          Point3D[] waypoints = new Point3D[] {new Point3D(), new Point3D()};
          waypoints[0].set(swingOverPlanarRegionsTrajectoryExpander.getExpandedWaypoints().get(0));
          waypoints[1].set(swingOverPlanarRegionsTrajectoryExpander.getExpandedWaypoints().get(1));
